requirements.txt
gunicorn==20.0.4
Django==2.2
pytz==2021.1
sqlparse==0.4.1
Pillow==8.1.1
django-filter==2.4.0
djangorestframework==3.12.2
sorl-thumbnail==12.7.0
reportlab==3.5.65
six==1.15.0
psycopg2-binary==2.8.5